% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{calcU}
\alias{calcU}
\title{Calculate,U, the harvest rate and maximum sustainable yield.}
\usage{
calcU(p, w)
}
\arguments{
\item{a}{scalar or vector of productivities.}

\item{the}{w parameter in the modified Beverton Holt.}
}
\value{
a plot of a model parameter by year and population.
}
\description{
Calculate,U, the harvest rate and maximum sustainable yield.
}
