% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDefaultPriors.R
\name{createDefaultPriors}
\alias{createDefaultPriors}
\title{create default model priors}
\usage{
createDefaultPriors(modParams = NULL)
}
\arguments{
\item{modParams}{a list with any priors that you would like to override.}
}
\value{
a list of priors.
}
\description{
create default model priors
}
\examples{
priors <- createDefaultPriors()
}
