% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGScode.R
\name{createJAGScode}
\alias{createJAGScode}
\title{creates the JAGS code for the model}
\usage{
createJAGScode(fixedObsError = NULL, smoltData = TRUE)
}
\description{
creates the JAGS code for the model
}
\examples{
bmod <- createJAGScode()
}
