% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJAGSdata.R
\name{createJAGSdata}
\alias{createJAGSdata}
\title{create a data list for the JAGS models}
\usage{
createJAGSdata(dataDir = "", dataType = "smolt", includeSmolt = TRUE,
  habType = "area")
}
\arguments{
\item{dataDir}{the directory of the file together.csv.}

\item{includeSmolt}{if TRUE, smolt data will be included}
}
\value{
a list with the data necessary to run the JAGS model.
}
\description{
create a data list for the JAGS models
}
\examples{
dat <- createJAGSprojectionData("data dir goes here")

}
