% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{expandAllData}
\alias{expandAllData}
\title{Expands data to include all years and observations for all data in bdat.}
\usage{
expandAllData(jdat)
}
\arguments{
\item{bdat}{a list with the data you would like to fill.}
}
\value{
a dataframe with all of the expanded data.
}
\description{
Expands data to include all years and observations for all data in bdat.
}
