% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{expandData}
\alias{expandData}
\title{Expands data to include all years and observations.}
\usage{
expandData(data, index, N)
}
\arguments{
\item{data}{the data you would like to fill.}

\item{index}{the index indicating which of the total values were observed.}

\item{N}{the total number of length of the vector to be created.}
}
\value{
a vector NAs of length N, with values at indices, index, replaced with data.
}
\description{
Expands data to include all years and observations.
}
