% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{fillGaps}
\alias{fillGaps}
\title{fills gaps in a series}
\usage{
fillGaps(vals, groups, returnEstType = FALSE)
}
\arguments{
\item{vals}{the values which you would like to fill.}

\item{groups}{the variable which defines groups.}

\item{returnEstType}{if TRUE, the function returns the method used to fill.}
}
\value{
a vector of values with gaps filled.
}
\description{
fills gaps in a series
}
