% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{fillGapsKern}
\alias{fillGapsKern}
\title{fills gaps in a series using neighbors}
\usage{
fillGapsKern(dat, param, fKern = c(rep(0.01, 1000), 1:3, 0, 3:1,
  rep(0.01, 1000)))
}
\arguments{
\item{dat}{the JAGS data object.}

\item{fKern}{the kernal used to fill.}

\item{var}{the variable to fill.}
}
\value{
a vector of values with gaps filled.
}
\description{
fills gaps in a series using neighbors
}
