% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPostDraws.R
\name{getPostDraws}
\alias{getPostDraws}
\title{get posterior draws from a model object}
\usage{
getPostDraws(x)
}
\arguments{
\item{runObject}{an object returned by runJAGSmodel.}
}
\value{
a list of model parameter posterior draws.
}
\description{
get posterior draws from a model object
}
\examples{
x <- getPostDraws(runObject)
}
