% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPosteriorSummary.R
\name{getPosteriorSummary}
\alias{getPosteriorSummary}
\title{get table summarizing the posterior distributions for the model parameters}
\usage{
getPosteriorSummary(runObject)
}
\arguments{
\item{runObject}{an object returned by runJAGSmodel.}
}
\value{
a table summarizing the model parameter posterior distributions.
}
\description{
get table summarizing the posterior distributions for the model parameters
}
\examples{
x <- getPosteriorSummary(runObject)
}
