% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRhat.R
\name{getRhat}
\alias{getRhat}
\title{a function to retrieve the Rhat values from the model object}
\usage{
getRhat(runObject)
}
\arguments{
\item{runObject}{an object returned by runJAGSmodel.}
}
\value{
the Rhat values for the model parameters.
}
\description{
a function to retrieve the Rhat values from the model object
}
\examples{
rr <- getRhat(runObject)
}
