% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotAge}
\alias{plotAge}
\title{Plot predicted vs observed age composition.}
\usage{
plotAge(runObj, p = 0.8, newWindow = FALSE)
}
\arguments{
\item{runObj}{an object returned by the runJAGSmodel function.}
}
\value{
a plot of predicted vs observed proportion age 4 escapement.
}
\description{
Plot predicted vs observed age composition.
}
