% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotEsc}
\alias{plotEsc}
\title{Plot predicted vs observed escapement.}
\usage{
plotEsc(runObj, p = 0.8, newWindow = FALSE, logScale = FALSE)
}
\arguments{
\item{runObj}{an object returned by the runJAGSmodel function.}
}
\value{
a plot of predicted vs observed escapement.
}
\description{
Plot predicted vs observed escapement.
}
