% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotParam}
\alias{plotParam}
\title{Plot a parameter by year and population.}
\usage{
plotParam()
}
\arguments{
\item{runObj}{an object returned by the runJAGSmodel function.}

\item{param}{either a character string with the name of the parameter or an array with the parameter posterior.}

\item{paramName}{optional parameter giving the full name of the model parameter.}

\item{plotScale}{optional parameters describing on what scale to plot the y-axis ("identity","log","logit").}

\item{a}{scalar or vector of productivities.}

\item{the}{w parameter in the modified Beverton Holt.}
}
\value{
a scalar or vector of U's.

a plot of a model parameter by year and population.
}
\description{
Plot a parameter by year and population.
}
