% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotPost}
\alias{plotPost}
\title{Plot posteriorsDists.}
\usage{
plotPost(parSims, parName = "par", mainText = "", nSims = 1000)
}
\arguments{
\item{parSims}{a array with samples from the posterior distribution.}

\item{nSims}{the number of simulations in each chain}
}
\value{
a plot of the posterior.
}
\description{
Plot posteriorsDists.
}
