% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXY.R
\name{plotXY}
\alias{plotXY}
\title{plots two variables against each other}
\usage{
plotXY(xx, yy, xUn = FALSE, yUn = FALSE, xLab = "", yLab = "",
  xLim = c(NA, NA), yLim = c(NA, NA), main = "", p = 0.8,
  pch = 16)
}
\arguments{
\item{xx}{the x value (either a vector or a matrix where rows represent the MCMC sims).}

\item{yy}{the y value (see above).}

\item{xUn}{if TRUE a credible interval will be plotted for the x variable.}

\item{yUn}{if TRUE a credible interval will be plotted for the y variable.}

\item{xLab}{the x label.}

\item{yLab}{the y label.}

\item{xLim}{the x limits}

\item{yLim}{the y limits}

\item{main}{the main text}

\item{p}{the probability used to define the credible interval.}

\item{pch}{the point type passed to plot.}
}
\value{
creates a plot.
}
\description{
plots two variables against each other
}
\examples{
plotXY(x,y)
}
