% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{Simulate data based on the fitted model}
\usage{
simulateData(runObj, years = 25, numSims = 10, obsDat = FALSE)
}
\arguments{
\item{runObj}{an object created by runJAGSmodel}

\item{years}{the number of years to project forward in time}

\item{numSims}{the number of forward simulations}

\item{obsDat}{should observation data be generated?}
}
\value{
predictions
}
\description{
Simulate data based on the fitted model
}
\examples{
simulate(runObj,years=10,numSims=100)
}
